+++ tests/project/test_brownie_config.py
@@ -118,6 +118,7 @@ def test_validate_cmd_settings():
         mnemonic: brownie
         account_keys_path: ../../
         fork: main
+        disable_cache: true
     """
     cmd_settings_dict = yaml.safe_load(cmd_settings)
     valid_dict = _validate_cmd_settings(cmd_settings_dict)