+++ brownie/_cli/networks.py
@@ -51,6 +51,7 @@ DEV_CMD_SETTINGS = (
     "accounts",
     "evm_version",
     "fork",
+    "disable_cache",
     "mnemonic",
     "account_keys_path",
     "block_time",
@@ -121,12 +122,15 @@ def _add(env, id_, *args):
     with _get_data_folder().joinpath("network-config.yaml").open() as fp:
         networks = yaml.safe_load(fp)
     if env.lower() == "development":
-        new = {
-            "name": args.pop("name"),
-            "id": id_,
-            "cmd": args.pop("cmd"),
-            "host": args.pop("host"),
-        }
+        try:
+            new = {
+                "name": args.pop("name"),
+                "id": id_,
+                "cmd": args.pop("cmd"),
+                "host": args.pop("host"),
+            }
+        except KeyError as exc:
+            raise ValueError(f"Missing field: {exc.args[0]}")
         if "timeout" in args:
             new["timeout"] = args.pop("timeout")
         new["cmd_settings"] = args
@@ -347,7 +351,7 @@ def _list_providers(verbose=False):
 
 def _parse_args(args):
     try:
-        args = dict(i.split("=") for i in args)
+        args = dict(i.split("=", maxsplit=1) for i in args)
     except ValueError:
         raise ValueError("Arguments must be given as key=value") from None