+++ CHANGELOG.md
@@ -6,12 +6,19 @@ This changelog format is based on [Keep a Changelog](https://keepachangelog.com/
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/eth-brownie/brownie)
+### Added
+- Allow to subscribe to events, with callbacks, using the corresponding contract instance.
+- Allow to retrieve events that have occurred between two blocks using the corresponding contract instance.
+- Allow listening for an event using the associated contract instance.
+
 ### Fixed
+- Fix download of dependencies with special characters in tag/version ([#1470](https://github.com/eth-brownie/brownie/pull/1470))
 - Deepcopy config so that disconnecting and reconnecting to networks works fully ([#1466](https://github.com/eth-brownie/brownie/pull/1466))
 
 ## [1.18.1](https://github.com/eth-brownie/brownie/tree/v1.18.1) - 2022-02-15
 ### Fixed
 - Correctly modify chain time when using `chain.mine` with Ganache v7 ([#1438](https://github.com/eth-brownie/brownie/pull/1438))
+-  Ensure `.env` file exists for installed packages ([#1504](https://github.com/eth-brownie/brownie/pull/1504))
 
 ## [1.18.0](https://github.com/eth-brownie/brownie/tree/v1.18.0) - 2022-02-14
 ### Added