+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow to subscribe to events, with callbacks, using the corresponding contract instance.
 - Allow to retrieve events that have occurred between two blocks using the corresponding contract instance.
 - Allow listening for an event using the associated contract instance.
+- Allow cache disabling when forking with ganache ([#1473](https://github.com/eth-brownie/brownie/pull/1473))
 
 ### Fixed
 - Fix download of dependencies with special characters in tag/version ([#1470](https://github.com/eth-brownie/brownie/pull/1470))